move fresh name creator into scala.reflect.internal.util
